Link support (#14)
Vtsls and typescript language server return three ranges with link support. 
1. The target range
2. The target selection range
3. The original selection range

Typescript language server returns the 2nd, vtsls returns the 1st. 

So if we have a file like 
```typescript
function foo(){
//...body
}
```
The first range would be the entire function while the 2nd is just the name of the function. 
This PR enables Link support which will return all three (in a slightly different format)