fix: handle line trimming (#34)
When we update a file, we need to tell the LSP so it can send diagnostics for it. To do this, we send a didChange notification. If the diagnostic responses are dropped, get_diagnostics will never return.


I noticed a case where appending whitespace to a file may cause lspMcp to send it a didChange notification but never receive a response. This PR prevents sending updates for these kinds of changes. These changes won't affect results of things like hover since they're just newlines on the end of the file.