fix: sendRequest bug if started isn't defined (#37)
Earlier, I modified the sendRequest to wait for the server to finish starting up instead of waiting for the server to begin starting up. This doesn't account for the possibility the server wasn't started when calling sendRequest and the promise not being defined. I think the logic elsewhere shouldn't allow this to happen, but here's a PR for it.