refactor: Move the file read into openFileContents (#39)
Currently callers of openFileContents are responsible for reading the file themselves. This PR moves that call into openFileContents, reducing duplicated code.

I'm going to add a lock on this function in my next PR and want to ensure that the file is read inside the function, not outside the function. 
While some functions may still read the file contents and pass that to openFileContents, the file watcher doesn't. This should still prevent concurrency bugs.