fix: Use previous result ID for unchanged responses (#49)
When we send a request for diagnostics, we attach a previous result ID to avoid duplication. If the new request has identical results to the previous, we get an unchanged response.
The logic was slightly wrong, I used returned the results for the current result ID (for which there aren't any) accidentally instead of the previous result ID.